+++ src/hooks/course/useCourseData.ts
@@ -1,0 +1,23 @@
+import {
+  useGetAdminEmailListByCourse,
+  useGetCourseCalendarStatus,
+} from '@/services/schedule/calendarQueries';
+
+export const useCourseData = ({ courseId }: { courseId: number }) => {
+  const {
+    data: adminList = [], //
+    isLoading: isAdminListLoading,
+  } = useGetAdminEmailListByCourse(courseId);
+
+  const {
+    data: courseCalendar, //
+    isLoading: isCourseCalendarLoading,
+  } = useGetCourseCalendarStatus(courseId);
+
+  return {
+    adminList,
+    isAdminListLoading,
+    courseCalendar,
+    isCourseCalendarLoading,
+  };
+};
